[None][refactor] PT _SymmetricMemory CUDA-IPC only; drop NCCL window path
Collapse the dual-backend selector added in 405d414a2a to a single
PyTorch _SymmetricMemory CUDA-IPC backend. The NCCL window / LSA
barrier code path is removed:

  - Move AsyncUlyssesOp + ulysses_a2a_async_{prepare,async} TORCH ops
    into a new file cpp/tensorrt_llm/thop/asyncUlyssesOp.cpp.
    alltoallOp.cpp is restored to its main-branch state (helix only).
  - Delete cpp/tensorrt_llm/kernels/lsaBarrierKernel.{h,cu} (~150
    lines): the NCCL device-API LSA barrier is no longer needed;
    handle->barrier(channel, timeout_ms) on the symm-mem handle is
    the cross-rank fence.
  - Drop ncclMemAlloc + ncclCommWindowRegister + ncclGetPeerDevicePointer
    + the 85-line bootstrapCeCommFromPg / NCCL_CTA_POLICY setenv block
    + the dev-id Store exchange + the cudaDeviceEnablePeerAccess loop.
    PT's empty_strided_p2p + rendezvous + handle->get_buffer_ptrs()
    replaces all of it.
  - Op signatures simplified: drop int[] group (no longer needed; we
    cache AsyncUlyssesOp by pg->getGroupName()).

Python side cleanups:

  - UlyssesAttention: drop the now-unused _group_ranks member;
    remove the explicit setup_async_pipeline() hook in favour of
    _ensure_async_state() invoked lazily inside _issue_async (which
    is already @compiler.disable'd).
  - modules/attention.py: drop the matching setup_async_pipeline()
    call site.

Microbench A/B (B200, B=2 S=6144 H=32 D=128, P=2/4/8 eager+graph)
under the dual-backend commit showed PT and NCCL tied within noise
across all 6 cells, so this collapse is performance-neutral.

End-to-end LTX-2 / WAN ws=8 e2e verification still pending.
